CHANGE ICONS ACTIONS

+++ app/views/criteres/index.php
@@ -29,16 +29,19 @@
                                     <span class="badge bg-info text-dark fs-6 px-3"><?= $c['coefficient'] ?></span>
                                 </td>
                                 <td class="text-end text-nowrap">
-                                    <a href="index.php?controller=critere&action=edit&id=<?= $c['id_critere'] ?>"
-                                       class="btn btn-sm btn-outline-primary me-1" title="Modifier">
-                                        <i class="bi bi-pencil"></i>
-                                    </a>
-                                    <a href="index.php?controller=critere&action=delete&id=<?= $c['id_critere'] ?>"
-                                       class="btn btn-sm btn-outline-danger"
-                                       title="Supprimer"
-                                       onclick="return confirm('Supprimer ce critere ?')">
-                                        <i class="bi bi-trash"></i>
-                                    </a>
+                                    <div class="dropdown-menu-container">
+                                        <button class="dropdown-menu-trigger" title="Actions">
+                                            <i class="bi bi-three-dots-vertical"></i>
+                                        </button>
+                                        <ul class="dropdown-menu-content">
+                                            <li><a href="index.php?controller=critere&action=edit&id=<?= $c['id_critere'] ?>" class="dropdown-menu-item">
+                                                <i class="bi bi-pencil"></i> Modifier
+                                            </a></li>
+                                            <li><a href="index.php?controller=critere&action=delete&id=<?= $c['id_critere'] ?>" class="dropdown-menu-item danger" data-confirm="Supprimer ce critere ?">
+                                                <i class="bi bi-trash"></i> Supprimer
+                                            </a></li>
+                                        </ul>
+                                    </div>
                                 </td>
                             </tr>
                         <?php endforeach; ?>
