+++ app/views/entreprises/index.php
@@ -29,16 +29,19 @@
                                 <td><?= htmlspecialchars($e['secteur_activite'] ?? '') ?></td>
                                 <td><?= htmlspecialchars($e['ville'] ?? '') ?></td>
                                 <td class="text-end text-nowrap">
-                                    <a href="index.php?controller=entreprise&action=edit&id=<?= $e['numero_entreprise'] ?>"
-                                       class="btn btn-sm btn-outline-primary me-1" title="Modifier">
-                                        <i class="bi bi-pencil"></i>
-                                    </a>
-                                    <a href="index.php?controller=entreprise&action=delete&id=<?= $e['numero_entreprise'] ?>"
-                                       class="btn btn-sm btn-outline-danger"
-                                       title="Supprimer"
-                                       onclick="return confirm('Supprimer cette entreprise ?')">
-                                        <i class="bi bi-trash"></i>
-                                    </a>
+                                    <div class="dropdown-menu-container">
+                                        <button class="dropdown-menu-trigger" title="Actions">
+                                            <i class="bi bi-three-dots-vertical"></i>
+                                        </button>
+                                        <ul class="dropdown-menu-content">
+                                            <li><a href="index.php?controller=entreprise&action=edit&id=<?= $e['numero_entreprise'] ?>" class="dropdown-menu-item">
+                                                <i class="bi bi-pencil"></i> Modifier
+                                            </a></li>
+                                            <li><a href="index.php?controller=entreprise&action=delete&id=<?= $e['numero_entreprise'] ?>" class="dropdown-menu-item danger" data-confirm="Supprimer cette entreprise ?">
+                                                <i class="bi bi-trash"></i> Supprimer
+                                            </a></li>
+                                        </ul>
+                                    </div>
                                 </td>
                             </tr>
                         <?php endforeach; ?>
